Separate out the animations to own files
